Revert "Feature: Bedtime calculator"
This reverts commit d5274bd9ea48196758be804b1455990656dc2f46

After toying with this for a while, its kind of stupid.  I didnt really see that at first, but theres a few things wrong with its whole implementation and concept.  Its supposed to let you know what ideal times you need to go to sleep to get the most out of sleep cycles, but going to bed at, say, 9:16pm, does NOT mean you will fall asleep at 9:16pm.  Some people it takes an hour or more to actually fall asleep, others, less than a minute.  So right off the bat, this would be flawed.  Its too bad.  It was a great idea, but it would be too hard to get this to function for everyone, correctly.